titanic_api/requirements.txt
--index-url $(PIP_EXTRA_INDEX_URL)
titanic-model-jd88
--extra-index-url https://pypi.org/simple/
pandas>=1.3.5,<1.4.0
fastapi>=0.64.0<1.0.0
uvicorn>=0.11.3,<0.12.0
loguru>=0.5.3,<0.6.0
python-multipart>=0.0.5,<0.1.0
pydantic>=1.8.2,<1.9.0
typing_extensions>=3.7.4,<3.8.0
